requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
APScheduler==3.6.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
datasets==2.19.1
DAWG-Python==0.7.2
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
docker-pycreds==0.4.0
docopt==0.6.2
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.14.0
frozenlist==1.4.1
fsspec==2023.9.2
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==1.0.0
grpcio==1.64.0
h11==0.13.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.1
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.2
numpy==1.22.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
pip==21.2.4
platformdirs==4.2.2
prompt-toolkit==3.0.43
protobuf==4.25.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn==1.6.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
pydotenv==0.0.7
Pygments==2.18.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
python-dateutil==2.9.0
python-dotenv==1.0.1
python-telegram-bot==13.13
pytz==2023.2
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==2.7.0
sentencepiece==0.2.0
sentry-sdk==2.3.1
seqeval==1.2.2
setproctitle==1.3.3
setuptools==58.0.4
simpletransformers==0.70.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
streamlit==1.35.0
sympy==1.12
telegram==0.0.1
tenacity==8.3.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.1
typing_extensions==4.12.0
tzdata==2023.1
tzlocal==5.2
urllib3==2.2.1
wandb==0.17.0
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.43.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.0
